style(structure): :art: improves on package organization
