+++ .idea/compiler.xml
@@ -8,12 +8,14 @@
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
         <module name="user_ms" />
+        <module name="whattodo_user_ms" />
       </profile>
     </annotationProcessing>
   </component>
   <component name="JavacSettings">
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
       <module name="user_ms" options="-parameters" />
+      <module name="whattodo_user_ms" options="-parameters" />
     </option>
   </component>
 </project>
