+++ src/main/java/br/com/user_ms/domain/adapters/UserAdapter.java
@@ -1,0 +1,8 @@
+package br.com.user_ms.domain.adapters;
+
+import br.com.user_ms.domain.entity.User;
+
+public interface UserAdapter {
+
+    User saveUser(User user);
+}
