+++ src/main/java/br/com/user_ms/domain/exceptions/UserCreateException.java
@@ -1,0 +1,7 @@
+package br.com.user_ms.domain.exceptions;
+
+public class UserCreateException extends RuntimeException{
+    public UserCreateException(String msg){
+        super(msg);
+    }
+}
