+++ src/main/java/br/com/user_ms/domain/port/UserRegisterPort.java
@@ -1,0 +1,9 @@
+package br.com.user_ms.domain.port;
+
+import br.com.user_ms.domain.port.model.UserRegisterRequest;
+import br.com.user_ms.domain.port.model.UserRegisterResponse;
+
+public interface UserRegisterPort {
+
+    UserRegisterResponse register(UserRegisterRequest userRegisterRequest);
+}
