+++ src/main/java/br/com/user_ms/domain/port/model/UserRegisterRequest.java
@@ -1,8 +1,6 @@
-package br.com.user_ms.useCase.dto;
+package br.com.user_ms.domain.port.model;
 
-import br.com.user_ms.useCase.entity.enums.Status;
-
-public class UserRegisterRequestDto {
+public class UserRegisterRequest {
 
     private String name;
     private String surname;
@@ -10,7 +8,7 @@ public class UserRegisterRequestDto {
     private String password;
     private Byte perfilImage;
 
-    public UserRegisterRequestDto(String name, String surname, String email, String password, Byte perfilImage) {
+    public UserRegisterRequest(String name, String surname, String email, String password, Byte perfilImage) {
         this.name = name;
         this.surname = surname;
         this.email = email;
