--- src/main/java/br/com/user_ms/useCase/excepions/UserCreateExeption.java
@@ -1,7 +1,0 @@
-package br.com.user_ms.useCase.excepions;
-
-public class UserCreateExeption extends RuntimeException{
-    public UserCreateExeption(String msg){
-        super(msg);
-    }
-}
