--- src/main/java/br/com/user_ms/useCase/port/UserRegisterPort.java
@@ -1,9 +1,0 @@
-package br.com.user_ms.useCase.port;
-
-import br.com.user_ms.useCase.dto.UserRegisterRequestDto;
-import br.com.user_ms.useCase.dto.UserRegisterResponseDto;
-
-public interface UserRegisterPort {
-
-    UserRegisterResponseDto register(UserRegisterRequestDto userRegisterRequestDto);
-}
